Added more unit tests

+++ notes-api/src/main/java/com/kchan/project/beyond/notes/controller/NotesController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.kchan.project.beyond.notes.dao.NotesRepository;
 import com.kchan.project.beyond.notes.dto.Note;
+import com.kchan.project.beyond.notes.exception.runtimes.NotFoundException;
 
 /*
  * Do we need to specify /application-json content?
@@ -47,25 +50,37 @@ public class NotesController {
 	@RequestMapping(method=RequestMethod.GET, value="/notes")
 	public @ResponseBody List<Note> readNotes(@RequestParam(required=false, value="query") String query) {
 		
-		if(query != null) logger.info(String.format("Query: [%s]", query));
+		logger.debug(String.format("Query: [%s]", query));
 		return (query ==  null) ? (List<Note>) this.repo.findAll() : this.repo.findByBodyContainingIgnoreCase(query); 
 	}
 	
 	@RequestMapping(method=RequestMethod.PUT, value="/notes")
-	public @ResponseBody String updateNote(@RequestBody Note note) {
+	public ResponseEntity<String> updateNote(@RequestBody Note note) {
 		
 		boolean noteFound = this.repo.exists(note.getId());
 		this.repo.save(note);
 		
-		return noteFound ? "Successful update" : "Note does not exist; saved note with a different ID";
+		logger.debug(String.format("\n\nUPDATE NOTE: %s", note.toString()));
+		
+		return noteFound ? 
+				new ResponseEntity<>("Update successful", HttpStatus.OK) : 
+				new ResponseEntity<>("Note does not exist; saved note with a different ID", HttpStatus.CREATED);
 	}
 	
+	/*
+	 * Synchronized to prevent possibility of Exception thrown during delete operation
+	 */
 	@RequestMapping(method=RequestMethod.DELETE, value="/notes/{id}")
-	public @ResponseBody String deleteNote(@PathVariable int id) {
+	public synchronized ResponseEntity<String> deleteNote(@PathVariable int id) {
 		
 		boolean noteFound = this.repo.exists(id);
+		
+		if(!noteFound) {
+			throw new NotFoundException("Note does not exist");
+		}
+		
 		this.repo.delete(id);
 		
-		return noteFound ? "Successful delete" : "Note does not exist";
+		return new ResponseEntity<>("Delete successful", HttpStatus.OK); 
 	}
 }
